Move IPoWBlockHeader to modAion from modAionBase
